+++ extensions-core/multi-stage-query/src/main/java/org/apache/druid/msq/logical/stages/OffsetLimitStage.java
@@ -19,46 +19,43 @@
 
 package org.apache.druid.msq.logical.stages;
 
-import org.apache.druid.frame.key.ClusterBy;
-import org.apache.druid.frame.key.KeyColumn;
-import org.apache.druid.msq.kernel.ShuffleSpec;
+import com.google.common.collect.ImmutableList;
+import org.apache.druid.msq.exec.StageProcessor;
 import org.apache.druid.msq.logical.LogicalInputSpec;
-import org.apache.druid.msq.querykit.QueryKitUtils;
-import org.apache.druid.msq.querykit.ShuffleSpecFactories;
+import org.apache.druid.msq.logical.StageMaker;
+import org.apache.druid.msq.querykit.common.OffsetLimitStageProcessor;
 import org.apache.druid.segment.column.RowSignature;
+import org.apache.druid.sql.calcite.planner.OffsetLimit;
 import org.apache.druid.sql.calcite.planner.querygen.DruidQueryGenerator.DruidNodeStack;
 
-import java.util.List;
-
-public class SortStage extends AbstractShuffleStage
+public class OffsetLimitStage extends AbstractFrameProcessorStage
 {
-  protected final List<KeyColumn> keyColumns;
+  protected final OffsetLimit offsetLimit;
 
-  public SortStage(LogicalStage inputStage, List<KeyColumn> keyColumns)
+  public OffsetLimitStage(LogicalStage inputStage, OffsetLimit offsetLimit)
   {
-    super(
-        QueryKitUtils.sortableSignature(inputStage.getLogicalRowSignature(), keyColumns),
-        LogicalInputSpec.of(inputStage)
-    );
-    this.keyColumns = keyColumns;
-  }
-
-  @Override
-  public RowSignature getLogicalRowSignature()
-  {
-    return inputSpecs.get(0).getRowSignature();
-  }
-
-  @Override
-  public ShuffleSpec buildShuffleSpec()
-  {
-    final ClusterBy clusterBy = new ClusterBy(keyColumns, 0);
-    return ShuffleSpecFactories.globalSortWithMaxPartitionCount(1).build(clusterBy, false);
+    super(inputStage.getRowSignature(), ImmutableList.of(LogicalInputSpec.of(inputStage)));
+    this.offsetLimit = offsetLimit;
   }
 
   @Override
   public LogicalStage extendWith(DruidNodeStack stack)
   {
     return null;
+  }
+
+  @Override
+  public StageProcessor<?, ?> buildStageProcessor(StageMaker stageMaker)
+  {
+    return new OffsetLimitStageProcessor(
+        offsetLimit.getOffset(),
+        offsetLimit.hasLimit() ? offsetLimit.getLimit() : null
+    );
+  }
+
+  @Override
+  public RowSignature getLogicalRowSignature()
+  {
+    return inputSpecs.get(0).getRowSignature();
   }
 }
